backend/requirements.txt
flask>=2.1
tensorflow>=2.10
pillow>=9.0
numpy>=1.23
flask-cors>=3.0
gunicorn>=20.1
